server/requirements.txt
gabriel-server==2.1.1
asyncio==3.4.3
grpcio-tools==1.44.0
protobuf==3.20.3
pyzmq